remove build and fetching toggles from endpoint
