+++ toggles/webapp/deploy.ts
@@ -1,9 +1,7 @@
-import { GetObjectCommand, S3Client } from '@aws-sdk/client-s3';
-import { Readable } from 'stream';
-import localToggles, { Toggle } from './toggles';
+import { S3Client } from '@aws-sdk/client-s3';
 import { TogglesResp } from '.';
-import { bucket, key } from './config';
-import { putTogglesObject } from './aws';
+import { getTogglesObject, putTogglesObject } from './aws';
+import localToggles, { Toggle } from './toggles';
 
 export const withDefaultValuesUnmodified = (
   from: Toggle[],
@@ -27,27 +25,8 @@ export const withDefaultValuesUnmodified = (
   return toggles;
 };
 
-// see: https://github.com/aws/aws-sdk-js-v3/issues/1877
-async function streamToString(stream: Readable): Promise<string> {
-  return new Promise((resolve, reject) => {
-    const chunks: Uint8Array[] = [];
-    stream.on('data', chunk => chunks.push(chunk));
-    stream.on('error', reject);
-    stream.on('end', () => resolve(Buffer.concat(chunks).toString('utf-8')));
-  });
-}
-
 export async function deploy(client: S3Client) {
-  const getObjectCommand = new GetObjectCommand({
-    Bucket: bucket,
-    Key: key,
-  });
-
-  const { Body: body } = await client.send(getObjectCommand);
-
-  const remoteToggles: TogglesResp = JSON.parse(
-    await streamToString(body as Readable)
-  );
+  const remoteToggles = await getTogglesObject(client);
 
   const togglesToDeploy = withDefaultValuesUnmodified(remoteToggles.toggles, [
     ...localToggles.toggles,
