+++ toggles/webapp/package.json
@@ -19,7 +19,6 @@
   },
   "dependencies": {
     "@aws-sdk/client-cloudfront": "^3.39.0",
-    "node-fetch": "2.6.2",
     "yargs": "^17.2.1"
   }
 }
